Adding myvector vector package to mpreduce and changing various cas commands.
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@15676 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ desktop/geogebra/awt/GeneralPath.java
@@ -120,7 +120,7 @@ public class GeneralPath extends geogebra.common.awt.GeneralPath implements geog
 	}
 	@Override
 	public void curveTo(float f, float g, float h, float i, float j, float k) {
-		impl.curveTo(f, g, h, i, j, k);
+		impl.curveTo((float)f, (float)g, (float)h, (float)i, (float)j, (float)k);
 		
 	}
 }
